requirements.txt
pyotp
cryptography
pillow
pyzbar
opencv-python
pystray
tkinter
cv2